requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
blas=1.0=openblas
brotli-python=1.0.9=py312h313beb8_8
bzip2=1.0.8=h80987f9_6
ca-certificates=2024.7.2=hca03da5_0
certifi=2024.8.30=py312hca03da5_0
charset-normalizer=3.3.2=pyhd3eb1b0_0
expat=2.6.3=h313beb8_0
filelock=3.13.1=py312hca03da5_0
freetype=2.12.1=h1192e45_0
giflib=5.2.1=h80987f9_3
idna=3.7=py312hca03da5_0
jinja2=3.1.4=py312hca03da5_0
jpeg=9e=h80987f9_3
lcms2=2.12=hba8e193_0
lerc=3.0=hc377ac9_0
libcxx=14.0.6=h848a8c0_0
libdeflate=1.17=h80987f9_1
libffi=3.4.4=hca03da5_1
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libjpeg-turbo=2.0.0=h1a28f6b_0
libopenblas=0.3.21=h269037a_0
libpng=1.6.39=h80987f9_0
libtiff=4.5.1=h313beb8_0
libwebp=1.3.2=ha3663a8_0
libwebp-base=1.3.2=h80987f9_0
llvm-openmp=14.0.6=hc6e5704_0
lz4-c=1.9.4=h313beb8_1
markupsafe=2.1.3=py312h80987f9_0
mpmath=1.3.0=py312hca03da5_0
ncurses=6.4=h313beb8_0
networkx=3.2.1=py312hca03da5_0
numpy=2.0.1=py312h7f4fdc5_1
numpy-base=2.0.1=py312he047099_1
openjpeg=2.5.2=h54b8e55_0
openssl=3.0.15=h80987f9_0
pillow=10.4.0=py312h80987f9_0
pip=24.2=py312hca03da5_0
pysocks=1.7.1=py312hca03da5_0
python=3.12.4=h99e199e_1
pytorch=2.5.0.dev20240909=py3.12_0
pyyaml=6.0.1=py312h80987f9_0
readline=8.2=h1a28f6b_0
requests=2.32.3=py312hca03da5_0
setuptools=72.1.0=py312hca03da5_0
sqlite=3.45.3=h80987f9_0
sympy=1.13.2=py312hca03da5_0
tk=8.6.14=h6ba3021_0
torchaudio=2.5.0.dev20240909=py312_cpu
torchvision=0.20.0.dev20240909=py312_cpu
typing_extensions=4.11.0=py312hca03da5_0
tzdata=2024a=h04d1e81_0
urllib3=2.2.2=py312hca03da5_0
wheel=0.43.0=py312hca03da5_0
xz=5.4.6=h80987f9_1
yaml=0.2.5=h1a28f6b_0
zlib=1.2.13=h18a0788_1
zstd=1.5.5=hd90d995_2
